Update collection.liquid
Updated the original code to handle drag and drop events. Imported AJAX references from google. Set the images on the collection pages to be dragable and modified the page header to be droppable. 

The handleDropEvent function takes the ID of the item being dragged, which is set to the ID of the associated product. It then utilizes the Shopify AJAX method of POST /cart/add.js to add the item to cart.

To acknowledge the item being added to the cart, the header switches to "Item Added to Cart" for 3 seconds, and then switches back to the original page header.

Finally we update the Cart Counter in the header with the appropriate number of Items in the cart. This is also accessed through the Shopify AJAX of getJSON /cart.js

To make the page header, and the product images available to jquery:
-Modified theme.liquid Layout to update the ID of the header to "header-wrap" and updated the cart item counter to "cart_span"
-Added a message box to the header.
-Modified product-grid-item.liquid Snippet Product image ID to the first available variable ID of the product.

These pages are also included under this branch for viewing.
Modified the