Added functionality for variations
Sometimes you just feel like an idiot... this was one of those moments. I wanted to add option for variations from the start but couldn't figure out how to get the products variations by passing the value through. After some forum surfing I realized I was going about it the wrong way. Instead of passing the Product ID to the Drag/Drop I needed to pass the product handle and then work from there.

The code has been updated to use the product handle that is passed from the product-grid-item.liquid snippet. It first checks if there is any variations of the product, if so it fills a drop down in the theme.liquid layout and displays it to the client. They then select the option they'd like and it then adds to cart.

If there are no variation it adds to cart in the normal way.

I also updated the client side transitions of the header to display the different options.